+++ clang-tools-extra/clang-tidy/bsl/BslTidyModule.cpp
@@ -39,6 +39,7 @@
 #include "LiteralsNoOctalCheck.h"
 #include "LiteralsUnsignedSuffixCheck.h"
 #include "LiteralsUserDefinedCheck.h"
+#include "NameCaseCheck.h"
 #include "NamePrefixesCheck.h"
 #include "NamespaceGlobalCheck.h"
 #include "NodiscardCheck.h"
@@ -128,6 +129,8 @@ public:
         "bsl-literals-digit-separator");
     CheckFactories.registerCheck<LiteralsUserDefinedCheck>(
         "bsl-literals-user-defined");
+    CheckFactories.registerCheck<NameCaseCheck>(
+        "bsl-name-case");
     CheckFactories.registerCheck<NamePrefixesCheck>(
         "bsl-name-prefixes");
     CheckFactories.registerCheck<NamespaceGlobalCheck>(